MVC pattern for Search
