+++ app/src/main/res/layout/activity_search.xml
@@ -62,7 +62,8 @@
         android:layout_height="wrap_content"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         tools:itemCount="5"
-        tools:listitem="@layout/track_view" />
+        tools:listitem="@layout/track_view"
+        />
 
     <TextView
         android:id="@+id/empty_list_problem_text"
